refactor: Footer의 복잡한 관심사 분리(사운드 핸들러, 전체화면 로직 분리)
